lots of refactoring, basic kvs implementation, boilerplate for dataset and request queue

+++ src/apify/storages/key_value_store.py
@@ -1,11 +1,15 @@
-from typing import Any, Dict, Optional, TypedDict, TypeVar, Union, overload
+from typing import Any, AsyncIterator, Dict, Optional, Tuple, TypeVar, Union, overload
 
 from apify_client import ApifyClientAsync
 from apify_client.clients import KeyValueStoreClientAsync
-from ..memory_storage.memory_storage import MemoryStorage
-from ..memory_storage.resource_clients.key_value_store import KeyValueStoreClient
 
+# from ..consts import StorageTypes
+from ..config import Configuration
+from ..memory_storage import MemoryStorage
+from ..memory_storage.resource_clients import KeyValueStoreClient
 from ._utils import _purge_default_storages
+
+# from .storage_manager import StorageManager
 
 T = TypeVar('T')
 # S = TypeVar('S', dict)
@@ -38,12 +42,14 @@ static async getInput<T = Dictionary | string | Buffer>(): Promise<T | null>
 
 
 class KeyValueStore:
+    """TODO: docs."""
+
     _id: str
     _name: Optional[str]
     _client: Union[KeyValueStoreClientAsync, KeyValueStoreClient]
-    _persist_state_event_started: bool = False
+    # _persist_state_event_started: bool = False
 
-    _cache: Dict[str, Dict]
+    # _cache: Dict[str, Dict]
     """Cache for persistent (auto-saved) values. When we try to set such value, the cache will be updated automatically."""
 
     def __init__(self, id: str, name: Optional[str], client: Union[ApifyClientAsync, MemoryStorage]) -> None:
@@ -53,47 +59,72 @@ class KeyValueStore:
         self._client = client.key_value_store(self._id)
 
     @classmethod
-    async def open(cls, store_id_or_name: str, client: Union[ApifyClientAsync, MemoryStorage]) -> 'KeyValueStore':
-        await _purge_default_storages(client)
+    async def open(cls, store_id_or_name: str, client: Union[ApifyClientAsync, MemoryStorage], config: Configuration) -> 'KeyValueStore':
+        """TODO: docs."""
+        if config.purge_on_start:
+            await _purge_default_storages(client)
 
-        id = store_id_or_name
-        name = None
         key_value_store_client = client.key_value_store(store_id_or_name)
-        if not await key_value_store_client.get():
+        key_value_store_info = await key_value_store_client.get()
+        if not key_value_store_info:
             key_value_store_info = await client.key_value_stores().get_or_create(name=store_id_or_name)
-            id = key_value_store_info['id']
-            name = key_value_store_info['name']
 
-        return KeyValueStore(id, name, client)
+        return KeyValueStore(key_value_store_info['id'], key_value_store_info['name'], client)
 
     @overload
-    async def get_value(self, key: str) -> Any: # TODO: This overload stuff seems useless?
+    async def get_value(self, key: str) -> Any:  # noqa: U100
+        """TODO: docs."""
         ...
+
     @overload
-    async def get_value(self, key: str, default_value: T) -> T:
+    async def get_value(self, key: str, default_value: T) -> T:  # noqa: U100
+        """TODO: docs."""
         ...
 
     async def get_value(self, key: str, default_value: Optional[T] = None) -> Optional[T]:
+        """TODO: docs."""
         record = await self._client.get_record(key)
         return record['value'] if record else default_value
 
     # async def get_auto_saved_value(self, key: str, default_value: Optional[S] = None) -> Optional[S]:
     #     pass
 
-    # async def for_each_key(self):
-    #     pass
+    def for_each_key(self, exclusive_start_key: Optional[str] = None) -> AsyncIterator[Tuple[Dict, int, int]]:
+        """TODO: docs."""
+        return self._for_each_key(exclusive_start_key)
+
+    async def _for_each_key(self, exclusive_start_key: Optional[str] = None, index: int = 0) -> AsyncIterator[Tuple[Dict, int, int]]:
+        """TODO: docs."""
+        list_keys = await self._client.list_keys(exclusive_start_key=exclusive_start_key, limit=1)
+        for item in list_keys['items']:
+            yield item, index, item['size']
+            index += 1
+
+        if list_keys['isTruncated']:  # TODO: Can we somehow simplify this with a recursion return? Possibly try AsyncGenerator or some Union[...]
+            async for i, idx, size in self._for_each_key(list_keys['nextExclusiveStartKey'], index):
+                yield i, idx, size
 
     # async def _ensure_persist_state_event(self):
     #     pass
 
     async def set_value(self, key: str, value: Optional[T], content_type: Optional[str] = None) -> None:
-        pass
+        """TODO: docs."""
+        if value is None:
+            return await self._client.delete_record(key)
 
-    async def _clear_cache(self) -> None: # TODO: Crawlee has this only for tests, so most probably not even needed
-        self._cache.clear()
+        # TODO: Why does crawlee run another stringify here?
+        # value = maybeStringify(value, optionsCopy)
+
+        return await self._client.set_record(key, value, content_type)
+
+    # async def _clear_cache(self) -> None:  # TODO: Crawlee has this only for tests, so most probably not even needed
+    #     self._cache.clear()
 
     async def drop(self) -> None:
+        """TODO: docs."""
         # await this.client.delete();
         # const manager = StorageManager.getManager(KeyValueStore, this.config);
         # manager.closeStorage(this);
         await self._client.delete()
+        # TODO: Resolve circular imports...
+        # await StorageManager.close_storage(StorageTypes.KEY_VALUE_STORE, self._id)
